Moving the retry logic to the observers so the sleep fires in the correct thread
