+++ .github/workflows/main.yml
@@ -33,7 +33,7 @@ jobs:
       env:
         COVERALLS_REPO_TOKEN: ${{ secrets.COVERALLS_REPO_TOKEN }}
       run: |
-        ./gradlew :ivi-sdk-java:test :ivi-sdk-java:jacocoTestReport :ivi-sdk-java:coverallsJacoco
+        ./gradlew :ivi-sdk-java:jacocoTestReport :ivi-sdk-java:coverallsJacoco
 
     - name: Build Snapshot
       if: github.ref == 'refs/heads/develop' && github.event_name == 'push'
