+++ ivi-sdk-java/src/main/java/games/mythical/ivi/sdk/client/AbstractIVIClient.java
@@ -7,7 +7,6 @@ import io.grpc.CallCredentials;
 import io.grpc.ManagedChannel;
 import io.grpc.Metadata;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.RandomUtils;
 import org.apache.commons.lang3.StringUtils;
 
 import java.util.concurrent.Executor;
@@ -23,11 +22,6 @@ public abstract class AbstractIVIClient {
     // gRPC settings
     protected final int keepAlive;
     protected ManagedChannel channel;
-
-    // retry settings
-    private boolean retry = false;
-    private int requestCount = 1;
-    private final int maxCount = 16;
 
     protected AbstractIVIClient() throws IVIException {
         if(StringUtils.isEmpty(IVIConfiguration.getEnvironmentId())) {
@@ -66,30 +60,5 @@ public abstract class AbstractIVIClient {
             @Override
             public void thisUsesUnstableApi() { }
         };
-    }
-
-    protected void sleepBetweenReconnects() {
-        if(!retry) {
-            retry = true;
-            return;
-        }
-
-        var sleepTimeMillis = (int) Math.pow(2, requestCount) + RandomUtils.nextInt(1, 1000);
-        try {
-            log.trace("Sleeping {} milliseconds before reconnect", sleepTimeMillis);
-            Thread.sleep(sleepTimeMillis);
-
-            if(requestCount < maxCount) {
-                requestCount++;
-            }
-        } catch (InterruptedException e) {
-            log.error("Retry interrupted, exiting...");
-            Thread.currentThread().interrupt();
-        }
-    }
-
-    protected void resetConnectionRetry() {
-        retry = false;
-        requestCount = 1;
     }
 }
