+++ ivi-sdk-java/src/main/java/games/mythical/ivi/sdk/client/observer/IVIItemObserver.java
@@ -6,33 +6,28 @@ import games.mythical.ivi.sdk.proto.common.item.ItemState;
 import games.mythical.ivi.sdk.proto.streams.item.ItemStatusConfirmRequest;
 import games.mythical.ivi.sdk.proto.streams.item.ItemStatusUpdate;
 import games.mythical.ivi.sdk.proto.streams.item.ItemStreamGrpc;
-import games.mythical.ivi.sdk.util.Procedure;
-import io.grpc.stub.StreamObserver;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.function.Consumer;
 
 @Slf4j
-public class IVIItemObserver implements StreamObserver<ItemStatusUpdate> {
+public class IVIItemObserver extends AbstractObserver<ItemStatusUpdate> {
     private final IVIItemExecutor IVIItemExecutor;
     private final ItemStreamGrpc.ItemStreamBlockingStub streamBlockingStub;
     private final Consumer<IVIItemObserver> resubscribe;
-    private final Procedure reset;
 
     public IVIItemObserver(IVIItemExecutor IVIItemExecutor,
                            ItemStreamGrpc.ItemStreamBlockingStub streamBlockingStub,
-                           Consumer<IVIItemObserver> resubscribe,
-                           Procedure reset) {
+                           Consumer<IVIItemObserver> resubscribe) {
         this.IVIItemExecutor = IVIItemExecutor;
         this.streamBlockingStub = streamBlockingStub;
         this.resubscribe = resubscribe;
-        this.reset = reset;
     }
 
     @Override
     public void onNext(ItemStatusUpdate message) {
         log.trace("ItemObserver.onNext for item: {}", message.getGameInventoryId());
-        reset.invoke();
+        resetConnectionRetry();
         try {
             IVIItemExecutor.updateItem(message.getGameInventoryId(),
                     message.getGameItemTypeId(),
@@ -51,12 +46,14 @@ public class IVIItemObserver implements StreamObserver<ItemStatusUpdate> {
     @Override
     public void onError(Throwable t) {
         log.error("ItemObserver.onError", t);
+        sleepBetweenReconnects();
         resubscribe.accept(this);
     }
 
     @Override
     public void onCompleted() {
         log.info("ItemObserver stream closed");
+        sleepBetweenReconnects();
         resubscribe.accept(this);
     }
