--- ivi-sdk-java/src/main/java/games/mythical/ivi/sdk/util/Procedure.java
@@ -1,6 +1,0 @@
-package games.mythical.ivi.sdk.util;
-
-@FunctionalInterface
-public interface Procedure {
-    void invoke();
-}
